+++ src/components/formComponents/ReadOnlyBox/index.js
@@ -1,19 +1,15 @@
-import {
-  Typography,
-  Box,
-  Button,
-  Collapse
-} from '@mui/material/index';
+import { Typography, Box, Button, Collapse } from '@mui/material/index';
 import VisibilityIcon from '@mui/icons-material/Visibility';
 import VisibilityOffIcon from '@mui/icons-material/VisibilityOff';
 import { useTheme } from '@mui/material/styles';
 import { UserContext } from 'context/user/index';
 import { useContext, useEffect, useState } from 'react';
 
-const ReadOnlyBox = ({ children, title }) => {
+const ReadOnlyBox = ({ children, title, alwaysOpen }) => {
   const theme = useTheme();
   const { advancedMode } = useContext(UserContext);
   const [isOpen, setIsOpen] = useState(false);
+  const actuallyOpen = alwaysOpen || isOpen;
 
   useEffect(() => {
     setIsOpen(advancedMode);
@@ -21,10 +17,15 @@ const ReadOnlyBox = ({ children, title }) => {
 
   return (
     <>
-      <Button color="primary" variant={isOpen ? "contained" : "outlined"} sx={{ paddingX: 1, paddingY: 0.5, minWidth: 0, display: 'inline-flex', gap: 1, alignItems: 'center' }} onClick={() => setIsOpen(!isOpen)}>
-        {isOpen ? <VisibilityIcon /> : <VisibilityOffIcon />} {title || 'Zusatzinformationen'}
+      <Button
+        color="primary"
+        variant={actuallyOpen ? 'contained' : 'outlined'}
+        sx={{ paddingX: 1, paddingY: 0.5, minWidth: 0, display: 'inline-flex', gap: 1, alignItems: 'center' }}
+        onClick={() => setIsOpen(!isOpen)}
+      >
+        {actuallyOpen ? <VisibilityIcon /> : <VisibilityOffIcon />} {title || 'Zusatzinformationen'}
       </Button>
-      <Collapse in={isOpen}>
+      <Collapse in={actuallyOpen}>
         <Box
           sx={{
             padding: theme.shape.paddingBoxMedium,
