+++ src/formConfigs/annahmen/rules/conditional/index.js
@@ -1,0 +1,5 @@
+const rules = {
+  conditionalField: { '==': [{ var: 'conditionalControlField' }, 'show'] }
+};
+
+export default rules;
