+++ src/formConfigs/mitarbeiter/CalculationUpdater/index.jsx
@@ -1,0 +1,10 @@
+import React, { useEffect } from 'react';
+import { useFormikContext } from 'formik';
+
+const StundensatzRechnerValueUpdater = () => {
+  const { values = {}, setFieldValue } = useFormikContext();
+
+  return <React.Fragment />;
+};
+
+export default StundensatzRechnerValueUpdater;
