+++ src/formConfigs/mitarbeiter/rules/conditional/index.js
@@ -1,0 +1,5 @@
+const rules = {
+  conditionalField: { '==': [{ var: 'conditionalControlField' }, 'show'] }
+};
+
+export default rules;
