+++ src/formConfigs/mitarbeiter/rules/validation/index.js
@@ -1,0 +1,10 @@
+const rules = {
+  email: [
+    {
+      message: 'Geben Sie eine korrekte E-Mail Adresse an.',
+      rule: 'email'
+    }
+  ]
+};
+
+export default rules;
