--- src/formConfigs/testForm/rules/conditional/index.js
@@ -1,5 +1,0 @@
-const rules = {
-  conditionalField: { '==': [{ var: 'conditionalControlField' }, 'show'] }
-};
-
-export default rules;
