--- src/formConfigs/testForm/rules/validation/index.js
@@ -1,10 +1,0 @@
-const rules = {
-  email: [
-    {
-      message: 'Geben Sie eine korrekte E-Mail Adresse an.',
-      rule: 'email'
-    }
-  ]
-};
-
-export default rules;
