+++ src/layout/MainLayout/Drawer/DrawerContent/Navigation/NavItem.js
@@ -14,7 +14,6 @@ const NavItem = ({ item }) => {
   // const { drawerOpen } = useSelector((state) => state.menu);
   const drawerOpen = true;
   const isActive = useMemo(() => {
-    item.matchingUrlRegexp && console.log('pathname.match(item.matchingUrlRegexp)', pathname.match(item.matchingUrlRegexp));
     return pathname === item.url || (item.matchingUrlRegexp && pathname.match(item.matchingUrlRegexp));
   }, [pathname, item.url, item.matchingUrlRegexp]);
