+++ src/routes/ProtectedRoute.js
@@ -6,7 +6,6 @@ import { CircularProgress, Stack } from '@mui/material';
 const ProtectedRoute = ({ children }) => {
   const { user, requestStatusCodes } = useContext(UserContext);
   const navigate = useNavigate();
-  console.log('userid', user.uid);
 
   const FullPageLoader = useCallback(
     () => (
@@ -22,13 +21,11 @@ const ProtectedRoute = ({ children }) => {
   }
 
   if (requestStatusCodes.loadingUser === 200) {
-    console.log('should navigate to login');
     navigate('/login');
     return <FullPageLoader />;
   }
 
   if (!requestStatusCodes.loadingUser || requestStatusCodes.loadingUser === 102) {
-    console.log('loading user');
     return <FullPageLoader />;
   }
